Split up streamflow plots into separate modules
